# src/main/resources/META-INF/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Beer API",
    "description": "The official Beer API!  Search for both beers and breweries.",
    "contact": {
      "name": "Eric Wittmann",
      "url": "http://www.apicur.io/",
      "email": "eric.wittmann@redhat.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/beers/{beerId}": {
      "get": {
        "summary": "Get Info About a Beer",
        "description": "Returns full information about a single beer.",
        "operationId": "getBeer",
        "responses": {
          "200": {
            "description": "A single beer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beer"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a Beer",
        "description": "Updates information about a single beer.",
        "operationId": "updateBeer",
        "requestBody": {
          "description": "Updated beer information.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Beer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The beer was updated."
          }
        }
      },
      "delete": {
        "summary": "Delete a Beer",
        "description": "Removes a single beer from the data set.",
        "operationId": "deleteBeer",
        "responses": {
          "204": {
            "description": "The beer was deleted."
          }
        }
      },
      "parameters": [
        {
          "name": "beerId",
          "in": "path",
          "description": "Unique ID of a beer.",
          "required": true,
          "schema": {
            "format": "int32",
            "type": "integer"
          }
        }
      ]
    },
    "/breweries": {
      "get": {
        "summary": "Get All Breweries",
        "description": "Returns a list of all the breweries.",
        "operationId": "listAllBreweries",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Optional - filter by state.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "in": "query",
            "description": "Optional - filter by city.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all breweries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Brewery"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a Brewery",
        "description": "Adds a single brewery to the data set.",
        "operationId": "addBrewery",
        "requestBody": {
          "description": "New brewery information.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brewery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Brewery successfully added."
          }
        }
      }
    },
    "/breweries/{breweryId}": {
      "get": {
        "summary": "Gets Info About a Brewery",
        "description": "Returns full information about a single brewery.",
        "operationId": "getBrewery",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Brewery"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a Brewery",
        "description": "Updates information about a single brewery.",
        "operationId": "updateBrewery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brewery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {}
        }
      },
      "delete": {
        "summary": "Delete a Brewery",
        "description": "Removes a single brewery from the data set.",
        "operationId": "deleteBrewery",
        "responses": {
          "204": {}
        }
      },
      "parameters": [
        {
          "name": "breweryId",
          "in": "path",
          "description": "Unique ID of a brewery.",
          "required": true,
          "schema": {
            "format": "int32",
            "type": "integer"
          }
        }
      ]
    },
    "/breweries/{breweryId}/beers": {
      "get": {
        "summary": "Gets Beers From a Brewery",
        "description": "Returns all of the beers made by the brewery.",
        "operationId": "listBreweryBeers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Beer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adds a Beer to the Brewery",
        "description": "Adds a single beer to the data set for this brewery.",
        "operationId": "addBeerToBrewery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Beer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {}
        }
      },
      "parameters": [
        {
          "name": "breweryId",
          "in": "path",
          "required": true,
          "schema": {
            "format": "int32",
            "type": "integer"
          }
        }
      ]
    },
    "/beers": {
      "get": {
        "summary": "Get All Beers",
        "description": "Returns all of the beers in the database.",
        "operationId": "listAllBeers",
        "parameters": [
          {
            "name": "style",
            "in": "query",
            "description": "Optional - filter by style (lager, ipa, etc).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ounces",
            "in": "query",
            "description": "Optional - filter by size.",
            "schema": {
              "format": "float",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All the beers!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Beer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a Beer",
        "description": "Adds a single beer to the dataset.",
        "operationId": "addBeer",
        "requestBody": {
          "description": "The beer to add to the data set.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Beer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The beer was added."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Beer": {
        "title": "Root Type for Beer",
        "description": "The root of the Beer type's schema.",
        "required": [
          "abv",
          "id",
          "name",
          "style",
          "breweryId",
          "ounces"
        ],
        "type": "object",
        "properties": {
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "abv": {
            "format": "double",
            "type": "number"
          },
          "ibu": {
            "format": "double",
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "style": {
            "type": "string"
          },
          "breweryId": {
            "format": "int32",
            "type": "integer"
          },
          "ounces": {
            "format": "double",
            "type": "number"
          }
        }
      },
      "Brewery": {
        "title": "Root Type for Brewery",
        "description": "The root of the Brewery type's schema.",
        "type": "object",
        "properties": {
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "beer",
      "description": ""
    },
    {
      "name": "brewery",
      "description": ""
    }
  ]
}